requirements.txt
Cython==0.29.23
easydict==1.7 
einops==0.3.0
json-tricks==3.15.5
kornia==0.2.2
matplotlib==3.3.4
numpy==1.19.5
opencv-python==3.4.1.15
pandas==1.1.5
Pillow==8.2.0
protobuf==3.17.3
pycocotools==2.0.2
pytorch-lightning==1.1.0
PyYAML==5.4.1
scikit-image==0.17.2
scikit-learn==0.22
scipy==1.5.4
seaborn==0.11.1
SimpleITK==2.1.1.2
tensorboard==2.5.0
tensorboardX==2.2
torch==1.3.1+cu100
torchvision==0.4.2+cu100
tornado==6.1
tqdm==4.61.1
yacs==0.1.8
yamlloader==1.1.0